Implement and use a connection listener class
